Revert "SDL_test_md5.h: fix MD5UINT4 type to be really 32 bits."
This reverts commit 8eeca8c7db323d82ec257db9886e2ba16c0c892a as it's potentially an ABI break.

+++ include/SDL_test_md5.h
@@ -65,7 +65,7 @@ extern "C" {
 /* ------------ Definitions --------- */
 
 /* typedef a 32-bit type */
-  typedef Uint32 MD5UINT4;
+  typedef unsigned long int MD5UINT4;
 
 /* Data structure for MD5 (Message-Digest) computation */
   typedef struct {
